backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pypdf==5.1.0
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
groq==0.13.1
pydantic-settings==2.7.0
numpy==2.0.2
